Copy the Gradle getLibs task to a separate file.
In the original file there are tasks depending on Kotlin. Gradle unable to download Kotlin because it requires Kotlin already downloaded in order just to read and understand tasks that will never be used in this session. In other words, in order to download Kotlin files Gradle requires these files to be already downloaded.